feat: add photo and styles to developers
