+++ components/homeDeveloper/homeDeveloper.module.css
@@ -1,0 +1,83 @@
+.item {
+  width: 250px;
+  position: relative;
+  object-fit: cover;
+}
+
+.text {
+  display: flex;
+  align-items: center;
+  justify-content: flex-start;
+  flex-direction: column;
+  gap: 10px;
+  padding: 18px;
+  position: relative;
+  height: 100%;
+  text-align: center;
+  border: 1px solid #0cb4f1;
+  border-radius: 5px;
+  z-index: 1;
+  backface-visibility: hidden;
+  transition: transform 2s cubic-bezier(0.2, 0.85, 0.4, 1.275);
+}
+
+.face {
+  z-index: 0;
+  position: absolute;
+  top: 0;
+  width: 100%;
+  height: 100%;
+  text-align: center;
+  border: 1px solid #0cb4a1;
+  border-radius: 5px;
+  backface-visibility: hidden;
+  transform: rotateY(-180deg);
+  transition: transform 2s cubic-bezier(0.2, 0.85, 0.4, 1.275);
+}
+
+.item:hover .face {
+  transform: rotateY(0deg);
+  z-index: 2;
+}
+
+.item:hover .text {
+  transform: rotateY(180deg);
+  z-index: -1;
+}
+
+.img {
+  object-fit: cover;
+}
+
+.name {
+  font-size: 20px;
+  opacity: 0.9;
+}
+
+.link {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: 5px;
+  color: #000;
+  opacity: 0.9;
+  transition: opacity 0.2s;
+}
+
+.linkWrapper {
+  position: absolute;
+  left: 0;
+  right: 0;
+  bottom: 0;
+  padding: 10px;
+  background: rgba(255, 255, 255, 0.3);
+}
+
+.link:hover {
+  opacity: 0.7;
+}
+
+.description {
+  align-self: flex-start;
+  text-align: start;
+}
