+++ components/index.ts
@@ -17,6 +17,7 @@ export { EndpointInput } from './graphQl/endpointInput/endpointInput';
 export { SdlInput } from './graphQl/sdlInput/sdlInput';
 export { HistoryList } from './historyList/historyList';
 export { HomeList } from './homeList/homeList';
+export { HomeDeveloper } from './homeDeveloper/homeDeveloper';
 
 export type { SchemaType } from './graphQl/graphQlDocumentation/graphQlDocumentation.props';
 export type { HeaderType } from './graphQl/headersEditor/headersEditor.props';
