+++ containers/home/home.module.css
@@ -15,16 +15,6 @@
   justify-content: center;
   gap: 20px;
   text-align: justify;
-  /* position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%); */
-}
-
-@media (max-width: 980px) {
-  .welcome {
-    width: 460px;
-  }
 }
 
 .title {
@@ -35,8 +25,7 @@
 .text {
   align-self: center;
   width: 700px;
-  font-size: 14px;
-  opacity: 0.7;
+  font-size: 16px;
 }
 
 .links {
